Refactor consul copying and add tag and local support
